feat(scheduler): refactoring intervals to be user friendly
- weekly interval times changing from minutes
to hour/minute string format
- excluded time interval times changing from
epoch to iso8601 string format
- rename from_minutes to from
- rename to_minutes to to
- rename from_unix to from
- rename to_unit to to

Closes #39

+++ dotcommonitor/client/model_scheduler.go
@@ -20,8 +20,8 @@ type WeeklyInterval struct {
 
 // DateTimeInterval ... struct for non-repeat date time interval
 type DateTimeInterval struct {
-	From int `json:"From,omitempty"` // UNIX time format
-	To   int `json:"To,omitempty"`   // UNIX time format
+	From int64 `json:"From,omitempty"` // UNIX time format
+	To   int64 `json:"To,omitempty"`   // UNIX time format
 }
 
 // AssignedTo ... struct for non-repeat date time interval
